backend/requirements.txt
fastapi
httpx
uvicorn
psycopg2-binary
asyncpg
pydantic
python-dotenv 